Fix running on pyenv and non-macOS environments.
There was an import for `mach_o_change` that only worked on a macOS
system Python environment, and this prevented virtualenvs from being
created when run elsewhere, or in a pyenv. This was needed to patch a
Python binary inside a virtualenv.

We now make this import conditional. If not found, the attempt to call
it later will still fail, but it's already being handled gracefully,
falling back to using `install_name_tool`, which is the common tool for
patching Python on a standard Python install.