Fix and improve support for modern virtualenvs.
Modern versions of virtualenv made some major changes to generated
scripts (for `pip`, `wheel`, etc.), symlinks, and introduced a
`pyvenv.cfg` that the `site.py` depends on. These all caused issues with
parallel versions of Python being installed in the same virtualenv.

This change redoes much of the virtualenv process, working around all
these changes. We now sort the provided Python versions to ensure a
stable order (and making it easier to use by consumers) and carefully
handle the management of all versioned and unversioned `python` binaries
and generated scripts, patching them to use the right interpreter.

All actual Python binaries are put into their own `.bin-<version>/`
directory, which allows us to place the interpreter and a `pyvenv.cfg`,
which is needed to establish a modern Python environment. The binary
directory is the first place checked, which is why we take this
approach. The equivalent `pythonX.Y` scripts in `bin/` are just wrappers
around these.

For Python 2.x, the `pyvenv.cfg` is renamed to a version-specific file,
and `site.py` patched appropriately.

pypy is supported, but cannot be installed alongside CPython, due to
incompatibilities. It's also important to note that two different
versions of pypy will share the same `site-packages` directory, which is
beyond our control.