Fix running Python scripts from a generated virtualenv.
virtualenv-multiver 2.0 added support for modern virtualenv's pyvenv.cfg
files by moving the pythonX.Y interpreters into special .bin-X.Y
directories, where the pyvenv.cfg could live. It then added a wrapper
script to forward calls there.

This didn't always work. As execve documents, only compiled programs,
not interpreted scripts, can be used as a shell interpreter. That didn't
stop it from working on some systems, but it was clearly broken on
others, resulting the script (e.g., pip) being executed as a bash script
and not a Python script.

At the time this was written, we couldn't use a symlink, as pyvenv.cfg
would be found in the wrong location. It appears that's fixed (perhaps
in a modern virtualenv). As such, we should be safe to use a symlink
now, ditching the wrapper and avoiding the issues.