Fix the generated path to the pyvenv.cfg symlink.
This tweaks the generation of the `pyvenv.cfg` symlink to ensure that
the symlink points to the correct path inside the virtualenv. The path
being generated incorrectly had the virtualenv name prefixed, resulting
in a broken symlink.

Testing Done:
Created a virtualenv and verified the path to ` pyvenv.cfg` was
correct.

Reviewed at https://reviews.reviewboard.org/r/13087/