Add arguments for controlling pip install behavior.
In a default setup, virtualenv-multiver no longer works correctly on
versions of Python prior to 3.8. The reason is that a modern
`virtualenv` will install the `pip` for the version of Python associated
with the `virtualenv` command, and that's too new these days for older
environments. So we need to explicitly download an appropriate version,
or not install one at all.

It's important that we give some control over this behavior, so we now
take arguments for controlling whether downloads occur and whether
`pip`, `setuptools`, and `wheel` get installed.

Along with this, we have options for `--quiet`/`--verbose`, `--help`,
and `--version`.

While here, some legacy unneeded code has been removed, to keep this a
bit more maintainable.

Testing Done:
Made some virtualenvs using this, verifying it downloaded appropriate
versions of the packages for each version of Python.

Tested all the options to make sure they did what they were supposed
to.

Reviewed at https://reviews.reviewboard.org/r/14539/